Fix schedule tests
