+++ tests/utils.py
@@ -21,6 +21,7 @@ from gramex.handlers import BaseHandler
 watch_info = []
 ws_info = []
 counters = Counter()
+slow = {'value': 0, 'started': False, 'max': 10}
 
 
 def args_as_json(handler):
@@ -153,9 +154,30 @@ def on_deleted(event):
     watch_info.append({'event': event, 'type': 'deleted'})
 
 
-def slow_count(var, count=1, delay=0.01):
-    for x in range(count):
-        info[var] = x
+def schedule_start():
+    counters['schedule-key'] += 1
+
+
+def schedule_key():
+    return '%d' % counters['schedule-key']
+
+
+def slow_count_start():
+    slow['started'] = True
+    slow['start'] = slow['value']
+
+
+def slow_count_check(delay=0.01):
+    time.sleep(delay)
+    assert slow['start'] < slow['value'] < slow['max'] - 1, 'Schedule runs in parallel'
+
+
+def slow_count(delay=0.01):
+    # This runs in a thread and increments slow['value'] every 10ms
+    while not slow['started']:
+        time.sleep(delay)
+    for x in range(slow['max']):
+        slow['value'] = x
         time.sleep(delay)
